Extract classes UserRepository and TaskRepository from PromptRepository
