+++ backend/oasst_backend/api/v1/frontend_messages.py
@@ -16,7 +16,7 @@ def get_message_by_frontend_id(
     """
     Get a message by its frontend ID.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message_by_frontend_message_id(message_id)
     return utils.prepare_message(message)
 
@@ -29,7 +29,7 @@ def get_conv_by_frontend_id(
     Get a conversation from the tree root and up to the message with given frontend ID.
     """
 
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message_by_frontend_message_id(message_id)
     messages = pr.fetch_message_conversation(message)
     return utils.prepare_conversation(messages)
@@ -43,7 +43,7 @@ def get_tree_by_frontend_id(
     Get all messages belonging to the same message tree.
     Message is identified by its frontend ID.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message_by_frontend_message_id(message_id)
     tree = pr.fetch_message_tree(message.message_tree_id)
     return utils.prepare_tree(tree, message.message_tree_id)
@@ -56,7 +56,7 @@ def get_children_by_frontend_id(
     """
     Get all messages belonging to the same message tree.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message_by_frontend_message_id(message_id)
     messages = pr.fetch_message_children(message.id)
     return utils.prepare_message_list(messages)
@@ -70,7 +70,7 @@ def get_descendants_by_frontend_id(
     Get a subtree which starts with this message.
     The message is identified by its frontend ID.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message_by_frontend_message_id(message_id)
     descendants = pr.fetch_message_descendants(message)
     return utils.prepare_tree(descendants, message.id)
@@ -84,7 +84,7 @@ def get_longest_conv_by_frontend_id(
     Get the longest conversation from the tree of the message.
     The message is identified by its frontend ID.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message_by_frontend_message_id(message_id)
     conv = pr.fetch_longest_conversation(message.message_tree_id)
     return utils.prepare_conversation(conv)
@@ -98,7 +98,7 @@ def get_max_children_by_frontend_id(
     Get message with the most children from the tree of the provided message.
     The message is identified by its frontend ID.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message_by_frontend_message_id(message_id)
     message, children = pr.fetch_message_with_max_children(message.message_tree_id)
     return utils.prepare_tree([message, *children], message.id)
