+++ backend/oasst_backend/api/v1/leaderboards.py
@@ -12,7 +12,7 @@ def get_assistant_leaderboard(
     db: Session = Depends(deps.get_db),
     api_client: ApiClient = Depends(deps.get_trusted_api_client),
 ):
-    pr = PromptRepository(db, api_client, None)
+    pr = PromptRepository(db, api_client)
     return pr.get_user_leaderboard(role="assistant")
 
 
@@ -21,5 +21,5 @@ def get_prompter_leaderboard(
     db: Session = Depends(deps.get_db),
     api_client: ApiClient = Depends(deps.get_trusted_api_client),
 ):
-    pr = PromptRepository(db, api_client, None)
+    pr = PromptRepository(db, api_client)
     return pr.get_user_leaderboard(role="prompter")
