+++ backend/oasst_backend/api/v1/messages.py
@@ -29,7 +29,7 @@ def query_messages(
     """
     Query messages.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     messages = pr.query_messages(
         username=username,
         api_client_id=api_client_id,
@@ -51,7 +51,7 @@ def get_message(
     """
     Get a message by its internal ID.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message(message_id)
     return utils.prepare_message(message)
 
@@ -64,7 +64,7 @@ def get_conv(
     Get a conversation from the tree root and up to the message with given internal ID.
     """
 
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     messages = pr.fetch_message_conversation(message_id)
     return utils.prepare_conversation(messages)
 
@@ -76,7 +76,7 @@ def get_tree(
     """
     Get all messages belonging to the same message tree.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message(message_id)
     tree = pr.fetch_message_tree(message.message_tree_id)
     return utils.prepare_tree(tree, message.message_tree_id)
@@ -89,7 +89,7 @@ def get_children(
     """
     Get all messages belonging to the same message tree.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     messages = pr.fetch_message_children(message_id)
     return utils.prepare_message_list(messages)
 
@@ -101,7 +101,7 @@ def get_descendants(
     """
     Get a subtree which starts with this message.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message(message_id)
     descendants = pr.fetch_message_descendants(message)
     return utils.prepare_tree(descendants, message.id)
@@ -114,7 +114,7 @@ def get_longest_conv(
     """
     Get the longest conversation from the tree of the message.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message(message_id)
     conv = pr.fetch_longest_conversation(message.message_tree_id)
     return utils.prepare_conversation(conv)
@@ -127,7 +127,7 @@ def get_max_children(
     """
     Get message with the most children from the tree of the provided message.
     """
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     message = pr.fetch_message(message_id)
     message, children = pr.fetch_message_with_max_children(message.message_tree_id)
     return utils.prepare_tree([message, *children], message.id)
@@ -137,5 +137,5 @@ def get_max_children(
 def mark_message_deleted(
     message_id: UUID, api_client: ApiClient = Depends(deps.get_trusted_api_client), db: Session = Depends(deps.get_db)
 ):
-    pr = PromptRepository(db, api_client, None)
+    pr = PromptRepository(db, api_client)
     pr.mark_messages_deleted(message_id)
