+++ backend/oasst_backend/api/v1/stats.py
@@ -13,5 +13,5 @@ def get_message_stats(
     db: Session = Depends(deps.get_db),
     api_client: ApiClient = Depends(deps.get_trusted_api_client),
 ):
-    pr = PromptRepository(db, api_client, None)
+    pr = PromptRepository(db, api_client)
     return pr.get_stats()
