+++ backend/oasst_backend/api/v1/tasks.py
@@ -190,9 +190,9 @@ def request_task(
     api_client = deps.api_auth(api_key, db)
 
     try:
-        pr = PromptRepository(db, api_client, request.user)
+        pr = PromptRepository(db, api_client, client_user=request.user)
         task, message_tree_id, parent_message_id = generate_task(request, pr)
-        pr.store_task(task, message_tree_id, parent_message_id, request.collective)
+        pr.task_repository.store_task(task, message_tree_id, parent_message_id, request.collective)
 
     except OasstError:
         raise
@@ -217,11 +217,11 @@ def tasks_acknowledge(
     api_client = deps.api_auth(api_key, db)
 
     try:
-        pr = PromptRepository(db, api_client, user=None)
+        pr = PromptRepository(db, api_client)
 
         # here we store the message id in the database for the task
         logger.info(f"Frontend acknowledges task {task_id=}, {ack_request=}.")
-        pr.bind_frontend_message_id(task_id=task_id, frontend_message_id=ack_request.message_id)
+        pr.task_repository.bind_frontend_message_id(task_id=task_id, frontend_message_id=ack_request.message_id)
 
     except OasstError:
         raise
@@ -245,8 +245,8 @@ def tasks_acknowledge_failure(
     try:
         logger.info(f"Frontend reports failure to implement task {task_id=}, {nack_request=}.")
         api_client = deps.api_auth(api_key, db)
-        pr = PromptRepository(db, api_client, user=None)
-        pr.acknowledge_task_failure(task_id)
+        pr = PromptRepository(db, api_client)
+        pr.task_repository.acknowledge_task_failure(task_id)
     except (KeyError, RuntimeError):
         logger.exception("Failed to not acknowledge task.")
         raise OasstError("Failed to not acknowledge task.", OasstErrorCode.TASK_NACK_FAILED)
@@ -265,7 +265,7 @@ def tasks_interaction(
     api_client = deps.api_auth(api_key, db)
 
     try:
-        pr = PromptRepository(db, api_client, user=interaction.user)
+        pr = PromptRepository(db, api_client, client_user=interaction.user)
 
         match type(interaction):
             case protocol_schema.TextReplyToMessage:
@@ -323,6 +323,6 @@ def close_collective_task(
     api_key: APIKey = Depends(deps.get_api_key),
 ):
     api_client = deps.api_auth(api_key, db)
-    pr = PromptRepository(db, api_client, user=None)
+    pr = PromptRepository(db, api_client)
     pr.close_task(close_task_request.message_id)
     return protocol_schema.TaskDone()
