+++ backend/oasst_backend/prompt_repository.py
@@ -8,8 +8,10 @@ from uuid import UUID, uuid4
 import oasst_backend.models.db_payload as db_payload
 from loguru import logger
 from oasst_backend.journal_writer import JournalWriter
-from oasst_backend.models import ApiClient, Message, MessageReaction, Task, TextLabels, User
+from oasst_backend.models import ApiClient, Message, MessageReaction, TextLabels, User
 from oasst_backend.models.payload_column_type import PayloadContainer
+from oasst_backend.task_repository import TaskRepository, validate_frontend_message_id
+from oasst_backend.user_repository import UserRepository
 from oasst_shared.exceptions import OasstError, OasstErrorCode
 from oasst_shared.schemas import protocol as protocol_schema
 from oasst_shared.schemas.protocol import LeaderboardStats, SystemStats
@@ -19,87 +21,26 @@ from starlette.status import HTTP_403_FORBIDDEN, HTTP_404_NOT_FOUND
 
 
 class PromptRepository:
-    def __init__(self, db: Session, api_client: ApiClient, user: Optional[protocol_schema.User]):
+    def __init__(
+        self,
+        db: Session,
+        api_client: ApiClient,
+        client_user: Optional[protocol_schema.User] = None,
+        user_repository: Optional[UserRepository] = None,
+        task_repository: Optional[TaskRepository] = None,
+    ):
         self.db = db
         self.api_client = api_client
-        self.user = self.lookup_user(user)
+        self.user_repository = user_repository or UserRepository(db, api_client)
+        self.user = self.user_repository.lookup_client_user(client_user, create_missing=True)
         self.user_id = self.user.id if self.user else None
+        self.task_repository = task_repository or TaskRepository(
+            db, api_client, client_user, user_repository=self.user_repository
+        )
         self.journal = JournalWriter(db, api_client, self.user)
 
-    def lookup_user(self, client_user: protocol_schema.User) -> Optional[User]:
-        if not client_user:
-            return None
-        user: User = (
-            self.db.query(User)
-            .filter(
-                User.api_client_id == self.api_client.id,
-                User.username == client_user.id,
-                User.auth_method == client_user.auth_method,
-            )
-            .first()
-        )
-        if user is None:
-            # user is unknown, create new record
-            user = User(
-                username=client_user.id,
-                display_name=client_user.display_name,
-                api_client_id=self.api_client.id,
-                auth_method=client_user.auth_method,
-            )
-            self.db.add(user)
-            self.db.commit()
-            self.db.refresh(user)
-        elif client_user.display_name and client_user.display_name != user.display_name:
-            # we found the user but the display name changed
-            user.display_name = client_user.display_name
-            self.db.add(user)
-            self.db.commit()
-        return user
-
-    def validate_frontend_message_id(self, message_id: str) -> None:
-        # TODO: Should it be replaced with fastapi/pydantic validation?
-        if not isinstance(message_id, str):
-            raise OasstError(
-                f"message_id must be string, not {type(message_id)}", OasstErrorCode.INVALID_FRONTEND_MESSAGE_ID
-            )
-        if not message_id:
-            raise OasstError("message_id must not be empty", OasstErrorCode.INVALID_FRONTEND_MESSAGE_ID)
-
-    def bind_frontend_message_id(self, task_id: UUID, frontend_message_id: str):
-        self.validate_frontend_message_id(frontend_message_id)
-
-        # find task
-        task: Task = self.db.query(Task).filter(Task.id == task_id, Task.api_client_id == self.api_client.id).first()
-        if task is None:
-            raise OasstError(f"Task for {task_id=} not found", OasstErrorCode.TASK_NOT_FOUND, HTTP_404_NOT_FOUND)
-        if task.expired:
-            raise OasstError("Task already expired.", OasstErrorCode.TASK_EXPIRED)
-        if task.done or task.ack is not None:
-            raise OasstError("Task already updated.", OasstErrorCode.TASK_ALREADY_UPDATED)
-
-        task.frontend_message_id = frontend_message_id
-        task.ack = True
-        # ToDo: check race-condition, transaction
-        self.db.add(task)
-        self.db.commit()
-
-    def acknowledge_task_failure(self, task_id):
-        # find task
-        task: Task = self.db.query(Task).filter(Task.id == task_id, Task.api_client_id == self.api_client.id).first()
-        if task is None:
-            raise OasstError(f"Task for {task_id=} not found", OasstErrorCode.TASK_NOT_FOUND, HTTP_404_NOT_FOUND)
-        if task.expired:
-            raise OasstError("Task already expired.", OasstErrorCode.TASK_EXPIRED)
-        if task.done or task.ack is not None:
-            raise OasstError("Task already updated.", OasstErrorCode.TASK_ALREADY_UPDATED)
-
-        task.ack = False
-        # ToDo: check race-condition, transaction
-        self.db.add(task)
-        self.db.commit()
-
     def fetch_message_by_frontend_message_id(self, frontend_message_id: str, fail_if_missing: bool = True) -> Message:
-        self.validate_frontend_message_id(frontend_message_id)
+        validate_frontend_message_id(frontend_message_id)
         message: Message = (
             self.db.query(Message)
             .filter(Message.api_client_id == self.api_client.id, Message.frontend_message_id == frontend_message_id)
@@ -113,20 +54,48 @@ class PromptRepository:
             )
         return message
 
-    def fetch_task_by_frontend_message_id(self, message_id: str) -> Task:
-        self.validate_frontend_message_id(message_id)
-        task = (
-            self.db.query(Task)
-            .filter(Task.api_client_id == self.api_client.id, Task.frontend_message_id == message_id)
-            .one_or_none()
+    def insert_message(
+        self,
+        *,
+        message_id: UUID,
+        frontend_message_id: str,
+        parent_id: UUID,
+        message_tree_id: UUID,
+        task_id: UUID,
+        role: str,
+        payload: db_payload.MessagePayload,
+        payload_type: str = None,
+        depth: int = 0,
+    ) -> Message:
+        if payload_type is None:
+            if payload is None:
+                payload_type = "null"
+            else:
+                payload_type = type(payload).__name__
+
+        message = Message(
+            id=message_id,
+            parent_id=parent_id,
+            message_tree_id=message_tree_id,
+            task_id=task_id,
+            user_id=self.user_id,
+            role=role,
+            frontend_message_id=frontend_message_id,
+            api_client_id=self.api_client.id,
+            payload_type=payload_type,
+            payload=PayloadContainer(payload=payload),
+            depth=depth,
         )
-        return task
+        self.db.add(message)
+        self.db.commit()
+        self.db.refresh(message)
+        return message
 
     def store_text_reply(self, text: str, frontend_message_id: str, user_frontend_message_id: str) -> Message:
-        self.validate_frontend_message_id(frontend_message_id)
-        self.validate_frontend_message_id(user_frontend_message_id)
+        validate_frontend_message_id(frontend_message_id)
+        validate_frontend_message_id(user_frontend_message_id)
 
-        task = self.fetch_task_by_frontend_message_id(frontend_message_id)
+        task = self.task_repository.fetch_task_by_frontend_message_id(frontend_message_id)
 
         if task is None:
             raise OasstError(f"Task for {frontend_message_id=} not found", OasstErrorCode.TASK_NOT_FOUND)
@@ -174,7 +143,7 @@ class PromptRepository:
     def store_rating(self, rating: protocol_schema.MessageRating) -> MessageReaction:
         message = self.fetch_message_by_frontend_message_id(rating.message_id, fail_if_missing=True)
 
-        task = self.fetch_task_by_frontend_message_id(rating.message_id)
+        task = self.task_repository.fetch_task_by_frontend_message_id(rating.message_id)
         task_payload: db_payload.RateSummaryPayload = task.payload.payload
         if type(task_payload) != db_payload.RateSummaryPayload:
             raise OasstError(
@@ -201,7 +170,7 @@ class PromptRepository:
 
     def store_ranking(self, ranking: protocol_schema.MessageRanking) -> MessageReaction:
         # fetch task
-        task = self.fetch_task_by_frontend_message_id(ranking.message_id)
+        task = self.task_repository.fetch_task_by_frontend_message_id(ranking.message_id)
         if not task.collective:
             task.done = True
             self.db.add(task)
@@ -254,142 +223,6 @@ class PromptRepository:
                     f"task payload type mismatch: {type(task_payload)=} != {db_payload.RankConversationRepliesPayload}",
                     OasstErrorCode.TASK_PAYLOAD_TYPE_MISMATCH,
                 )
-
-    def store_task(
-        self,
-        task: protocol_schema.Task,
-        message_tree_id: UUID = None,
-        parent_message_id: UUID = None,
-        collective: bool = False,
-    ) -> Task:
-        payload: db_payload.TaskPayload
-        match type(task):
-            case protocol_schema.SummarizeStoryTask:
-                payload = db_payload.SummarizationStoryPayload(story=task.story)
-
-            case protocol_schema.RateSummaryTask:
-                payload = db_payload.RateSummaryPayload(
-                    full_text=task.full_text, summary=task.summary, scale=task.scale
-                )
-
-            case protocol_schema.InitialPromptTask:
-                payload = db_payload.InitialPromptPayload(hint=task.hint)
-
-            case protocol_schema.PrompterReplyTask:
-                payload = db_payload.PrompterReplyPayload(conversation=task.conversation, hint=task.hint)
-
-            case protocol_schema.AssistantReplyTask:
-                payload = db_payload.AssistantReplyPayload(type=task.type, conversation=task.conversation)
-
-            case protocol_schema.RankInitialPromptsTask:
-                payload = db_payload.RankInitialPromptsPayload(type=task.type, prompts=task.prompts)
-
-            case protocol_schema.RankPrompterRepliesTask:
-                payload = db_payload.RankPrompterRepliesPayload(
-                    type=task.type, conversation=task.conversation, replies=task.replies
-                )
-
-            case protocol_schema.RankAssistantRepliesTask:
-                payload = db_payload.RankAssistantRepliesPayload(
-                    type=task.type, conversation=task.conversation, replies=task.replies
-                )
-
-            case protocol_schema.LabelInitialPromptTask:
-                payload = db_payload.LabelInitialPromptPayload(
-                    type=task.type, message_id=task.message_id, prompt=task.prompt, valid_labels=task.valid_labels
-                )
-
-            case protocol_schema.LabelPrompterReplyTask:
-                payload = db_payload.LabelPrompterReplyPayload(
-                    type=task.type,
-                    message_id=task.message_id,
-                    conversation=task.conversation,
-                    reply=task.reply,
-                    valid_labels=task.valid_labels,
-                )
-
-            case protocol_schema.LabelAssistantReplyTask:
-                payload = db_payload.LabelAssistantReplyPayload(
-                    type=task.type,
-                    message_id=task.message_id,
-                    conversation=task.conversation,
-                    reply=task.reply,
-                    valid_labels=task.valid_labels,
-                )
-
-            case _:
-                raise OasstError(f"Invalid task type: {type(task)=}", OasstErrorCode.INVALID_TASK_TYPE)
-
-        task = self.insert_task(
-            payload=payload,
-            id=task.id,
-            message_tree_id=message_tree_id,
-            parent_message_id=parent_message_id,
-            collective=collective,
-        )
-        assert task.id == task.id
-        return task
-
-    def insert_task(
-        self,
-        payload: db_payload.TaskPayload,
-        id: UUID = None,
-        message_tree_id: UUID = None,
-        parent_message_id: UUID = None,
-        collective: bool = False,
-    ) -> Task:
-        c = PayloadContainer(payload=payload)
-        task = Task(
-            id=id,
-            user_id=self.user_id,
-            payload_type=type(payload).__name__,
-            payload=c,
-            api_client_id=self.api_client.id,
-            message_tree_id=message_tree_id,
-            parent_message_id=parent_message_id,
-            collective=collective,
-        )
-        self.db.add(task)
-        self.db.commit()
-        self.db.refresh(task)
-        return task
-
-    def insert_message(
-        self,
-        *,
-        message_id: UUID,
-        frontend_message_id: str,
-        parent_id: UUID,
-        message_tree_id: UUID,
-        task_id: UUID,
-        role: str,
-        payload: db_payload.MessagePayload,
-        payload_type: str = None,
-        depth: int = 0,
-    ) -> Message:
-        if payload_type is None:
-            if payload is None:
-                payload_type = "null"
-            else:
-                payload_type = type(payload).__name__
-
-        message = Message(
-            id=message_id,
-            parent_id=parent_id,
-            message_tree_id=message_tree_id,
-            task_id=task_id,
-            user_id=self.user_id,
-            role=role,
-            frontend_message_id=frontend_message_id,
-            api_client_id=self.api_client.id,
-            payload_type=payload_type,
-            payload=PayloadContainer(payload=payload),
-            depth=depth,
-        )
-        self.db.add(message)
-        self.db.commit()
-        self.db.refresh(message)
-        return message
 
     def insert_reaction(self, task_id: UUID, payload: db_payload.ReactionPayload) -> MessageReaction:
         if self.user_id is None:
@@ -519,8 +352,8 @@ class PromptRepository:
         """
         Mark task as done. No further messages will be accepted for this task.
         """
-        self.validate_frontend_message_id(frontend_message_id)
-        task = self.fetch_task_by_frontend_message_id(frontend_message_id)
+        validate_frontend_message_id(frontend_message_id)
+        task = self.task_repository.fetch_task_by_frontend_message_id(frontend_message_id)
 
         if not task:
             raise OasstError(
